pyproject.toml
[tool.poetry]
name = "exam"
version = "0.1.0"
description = ""
authors = ["dgkatziouras <dimitrios.gkatziouras@amadeus.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
kivy = "2.2.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
